Prepare atm_recover_large_step_variables_work for OpenACC porting
This includes splitting a long do loop on iCell so an if condition on
rk_step is outside the do loop. Though this condition shouldn't lead to
warp divergence, it can still be helpful to do branch evaluation on the
device best suited for it (the CPU). Since there was also limited data
re-use in the long do loop, it may help to reduce kernel launch
overhead.

Also line up whitespace so the bdyMaskCell if condition is apparent and
clean up some lines that end with whitespace.