Add the environment helper scripts for testing
In order to run test scripts outside of a testing framework, handling of
the environment setup should not be solely dependent on running within a
dedicated test framework. This has the added benefit of
compartmentalizing the duties of environment and dependency solving from
running the tests.

These environment scripts allow for the selection of a particular
environment with the default being the fqdn of the current host. From
there, arguments are routed using standard POSIX-sh to a respective
script. In the case of Derecho (applicable to any system using lmod) all
subsequent argument are treated as modules to load into the current
session.

The hostenv.sh script relies on one "argument" $AS_HOST being passed in
via variable setting to facilitate selection.

The helpers.sh script provides convenience features for substring
checking in sh, delayed environment variable expansion via eval, and
quick banner creation.

The derecho.sh script is included as the first supported environment.