Using SCOTCH_Num as integer type
PT-Scotch can be built with 64-bit integers or 32-bit integers. This is
determined from build-time options and the generic integer type SCOTCH_Num
covers either option. This commit replaces the type of certain integer
variables, in the MPAS interface to PT-Scotch, to use SCOTCH_Num to remain
consistent with the external library.

Note that MPAS doesn't yet support 64-bit integers, and variables of type
SCOTCH_Num are typecast into 32-bit integers in framework/mpas_block_decomp.F.
And in order for MPAS to correctly work with PT-Scotch, PT-Scotch must not be
built to use 64-bit integers.